third_party/blink/web_tests/platform/generic/fast/speechsynthesis/speech-synthesis-voices-expected.txt

This tests that we can get synthesizer voices on the Mac

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS voiceCount > 0 is true
PASS foundEnglishVoice is true
PASS foundDefaultVoice is true
PASS successfullyParsed is true

TEST COMPLETE
